--- Content.Client/Labels/EntitySystems/LabelSystem.cs
@@ -1,7 +1,0 @@
-using Content.Shared.Labels.EntitySystems;
-
-namespace Content.Client.Labels;
-
-public sealed partial class LabelSystem : SharedLabelSystem
-{
-}
