+++ Content.Server/Forensics/Systems/ForensicPadSystem.cs
@@ -1,4 +1,3 @@
-using Content.Server.Labels;
 using Content.Server.Popups;
 using Content.Shared.DoAfter;
 using Content.Shared.Examine;
@@ -7,6 +6,7 @@ using Content.Shared.Forensics.Components;
 using Content.Shared.IdentityManagement;
 using Content.Shared.Interaction;
 using Content.Shared.Inventory;
+using Content.Shared.Labels.EntitySystems;
 
 namespace Content.Server.Forensics
 {
