+++ Content.Server/Salvage/SalvageSystem.cs
@@ -16,7 +16,7 @@ using Content.Shared.Construction.EntitySystems;
 using Robust.Shared.Audio.Systems;
 using Robust.Shared.Map.Components;
 using Robust.Shared.Timing;
-using Content.Server.Labels;
+using Content.Shared.Labels.EntitySystems;
 using Robust.Shared.EntitySerialization.Systems;
 
 namespace Content.Server.Salvage
