+++ Content.Server/Shuttles/Commands/FTLDiskCommand.cs
@@ -1,8 +1,8 @@
 using Content.Server.Administration;
-using Content.Server.Labels;
 using Content.Shared.Administration;
 using Content.Shared.Hands.Components;
 using Content.Shared.Hands.EntitySystems;
+using Content.Shared.Labels.EntitySystems;
 using Content.Shared.Shuttles.Components;
 using Content.Shared.Storage;
 using Content.Shared.Storage.EntitySystems;
