+++ Content.Shared/Labels/Components/LabelComponent.cs
@@ -1,3 +1,4 @@
+using Content.Shared.Labels.EntitySystems;
 using Robust.Shared.GameStates;
 
 namespace Content.Shared.Labels.Components;
@@ -6,6 +7,7 @@ namespace Content.Shared.Labels.Components;
 /// Makes entities have a label in their name. Labels are normally given by <see cref="HandLabelerComponent"/>
 /// </summary>
 [RegisterComponent, NetworkedComponent, AutoGenerateComponentState]
+[Access(typeof(LabelSystem))]
 public sealed partial class LabelComponent : Component
 {
     /// <summary>
